backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
Markdown==3.4.1
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
